Use appropriate style for serialized strings
When psych parses YAML with string in the folded style as:

    a: >
      some
      inline
      content

and serializes it back to YAML,
then it renders the string in the literal style instead:

    a: |
      some inline content

This patch fixes this behaviour. When a string doesn't contain newlines
(blank line at the end is ignored) and it's not necessary to be quoted,
then it will use plain style or folded style according to its length and
the `line_width` option.

    # line_width = 100
    a: some inline content

    # line_width = 11
    a: >
      some inline
      content

This change was originally proposed by @Iazel.