Make Psych::Emitter store the IO object in an instance variable
The IO object given to Psych::Emitter#initialize is saved inside
LibYAML, yaml_emitter_t. This is a problem because Ruby's GC marker
can't find it, and can lead to use after free.

So, store the IO object in an instance variable and fetch from it every
time the write handler is called.

The segmentation fault can be reproduced with the following snippet:

  emitter = Psych::Emitter.new(open("/dev/null", "w"))
  GC.start
  emitter.start_stream(Psych::Parser::UTF16BE) # make it write something
  emitter.end_stream