Preserve time zone offset when deserializing times
Fix support of time zone offsets. Now the time zone offset
will be preserved when times are deserialized from YAML with
a specified offset. Previously, times other than UTC ("Z")
where converted to local time.

The test cases use two different offsets to account for the
possibility of an erratic success when the tests are run in
the corresponding time zone.