Parse.load with an IO/File which has a non-YAML allowed encoding will just set the encoding to UTF-8 and hope for the best.  This appears to be how libyaml works.
This issue was noticed in yaml/store because it extends pstore which will
create an IO in read_only mode as:

```ruby
RD_ACCESS = {mode: IO::RDONLY | IO::BINARY, encoding: Encoding::ASCII_8BIT}
```

The data in the test case I was debugging was in fact UTF-8 data and MRI was
happy to take this 8bit IO and pretend it is UTF-8.  Form-fitting ftw.