Move the responsibility of safe loading to Node
This change moves the responsibility of safely loading a YAML tree from
an internal detail of `Psych.safe_load` into `Psych::Nodes::Node`.

The advantage of this is that you can now more easily safely load trees
that you're working on without replicating the safety constraints of
`Psych.safe_load`. The delegation of responsibilities also now matches
for safely and unsafely loading a tree.